requirements.txt
appdirs==1.4.4
certifi==2020.6.20
cffi==1.14.3
cfgv==3.2.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1.1
distlib==0.3.1
Django==2.2.8
djangorestframework==3.11.2
drf-yasg==1.17.1
filelock==3.0.12
identify==1.5.6
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
nodeenv==1.5.0
openapi-codec==1.3.2
packaging==20.4
pre-commit==2.7.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.0.35
